Refactor to ColorModeProvider
