+++ frontend/src/Components/App.tsx
@@ -1,4 +1,5 @@
 import classes from './App.module.css';
+import { ColorModeProvider } from './ColorModeProvider/ColorModeProvider';
 import { PageSelector } from './PageSelector';
 import { TouchDetector } from './TouchDetector';
 import { WebSocketProvider } from './WebSocket';
@@ -7,7 +8,9 @@ export const App = () => (
   <div class={classes.app}>
     <WebSocketProvider>
       <TouchDetector>
-        <PageSelector />
+        <ColorModeProvider>
+          <PageSelector />
+        </ColorModeProvider>
       </TouchDetector>
     </WebSocketProvider>
   </div>
